Extract process_input/output to a mixin
