+++ lib/floe.rb
@@ -25,6 +25,7 @@ require_relative "floe/workflow/runner/podman"
 require_relative "floe/workflow/state"
 require_relative "floe/workflow/states/choice"
 require_relative "floe/workflow/states/fail"
+require_relative "floe/workflow/states/input_output_mixin"
 require_relative "floe/workflow/states/map"
 require_relative "floe/workflow/states/non_terminal_mixin"
 require_relative "floe/workflow/states/parallel"
